requirements.txt
langchain-community
python-dotenv
fastapi
pydantic
langgraph
langchain-core
langchain
langchain-tavily
langchain_openai
tavily-python
amadeus